I did  package and structure in the code
