+++ flaskblog/models.py
@@ -1,0 +1,29 @@
+from flaskblog import db
+from datetime import datetime
+from sqlalchemy import Column, Integer, String
+from sqlalchemy import ForeignKey
+from sqlalchemy.orm import relationship
+from sqlalchemy import Text
+
+
+class User(db.Model):
+    id = Column(Integer, primary_key=True)
+    username = Column(String(20), unique=True, nullable=False)
+    Email = Column(String(120), unique=True, nullable=False)
+    image_file = Column(String(20), nullable=False, default="default.jpg")
+    password = Column(String(60), nullable=False)
+    posts = relationship('Post', backref='author', lazy=True)
+
+    def __repr__(self):
+        return f"user('{self.username}','{self.Email}','{self.image_file}')"
+
+
+class Post(db.Model):
+    id = Column(Integer, primary_key=True)
+    title = Column(String(100), nullable=False)
+   # date_posted = Column(datetime, nullable=False, default=datetime.utcnow)
+    content = Column(Text, nullable=False)
+    user_id = Column(Integer, ForeignKey('user.id'), nullable=False)
+
+    def __repr__(self):
+        return f"Post('{self.title}','{self.date_posted}'"
